// web/src/api/generated/models/StoryKind.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const StoryKind = {
    MEMORY: 'MEMORY',
    HEART_MOMENT: 'HEART_MOMENT',
    MILESTONE: 'MILESTONE'
} as const;
export type StoryKind = typeof StoryKind[keyof typeof StoryKind];


export function instanceOfStoryKind(value: any): boolean {
    for (const key in StoryKind) {
        if (Object.prototype.hasOwnProperty.call(StoryKind, key)) {
            if (StoryKind[key as keyof typeof StoryKind] === value) {
                return true;
            }
        }
    }
    return false;
}

export function StoryKindFromJSON(json: any): StoryKind {
    return StoryKindFromJSONTyped(json, false);
}

export function StoryKindFromJSONTyped(json: any, ignoreDiscriminator: boolean): StoryKind {
    return json as StoryKind;
}

export function StoryKindToJSON(value?: StoryKind | null): any {
    return value as any;
}

export function StoryKindToJSONTyped(value: any, ignoreDiscriminator: boolean): StoryKind {
    return value as StoryKind;
}